requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.30.1=pypi_0
aiofiles=23.2.1=pypi_0
aiohttp=3.9.1=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.3.0=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.1.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
async-lru=2.0.4=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
babel=2.13.1=pypi_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
beautifulsoup4=4.12.2=pypi_0
bertviz=1.4.0=pypi_0
bitsandbytes=0.41.3.post2=pypi_0
blas=1.0=openblas
bleach=6.1.0=pypi_0
blessed=1.20.0=pypi_0
boto3=1.34.64=pypi_0
botocore=1.34.64=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.11.17=hbcca054_0
cachetools=5.4.0=pypi_0
certifi=2023.11.17=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
datasets=2.16.0=pypi_0
debugpy=1.6.7=py310h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
deepspeed=0.12.6=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.7=pypi_0
distro=1.9.0=pypi_0
dnspython=2.6.1=pypi_0
docstring-parser=0.15=pypi_0
einops=0.7.0=pypi_0
email-validator=2.1.1=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
evaluate=0.4.1=pypi_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
fastapi=0.111.0=pypi_0
fastapi-cli=0.0.4=pypi_0
fastjsonschema=2.19.0=pypi_0
ffmpy=0.3.2=pypi_0
filelock=3.13.1=pypi_0
flash-attn=2.5.8=pypi_0
fonttools=4.50.0=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2023.10.0=pypi_0
gpustat=1.1.1=pypi_0
gradio=4.39.0=pypi_0
gradio-client=1.1.1=pypi_0
h11=0.14.0=pypi_0
hjson=3.1.0=pypi_0
httpcore=1.0.2=pypi_0
httptools=0.6.1=pypi_0
httpx=0.26.0=pypi_0
huggingface-hub=0.24.2=pypi_0
idna=3.4=pypi_0
importlib-resources=6.4.0=pypi_0
inflect=7.2.1=pypi_0
ipykernel=6.26.0=pyhf8b6a83_0
ipython=8.17.2=pyh41d4057_0
ipywidgets=8.1.1=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jieba=0.42.1=pypi_0
jinja2=3.1.2=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.3.2=pypi_0
json5=0.9.14=pypi_0
jsonlines=2.0.0=pyhd3eb1b0_2
jsonpointer=2.4=pypi_0
jsonschema=4.20.0=pypi_0
jsonschema-specifications=2023.11.2=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.6.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-events=0.9.0=pypi_0
jupyter-lsp=2.2.1=pypi_0
jupyter-server=2.11.1=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyter_core=5.5.0=py310hff52083_0
jupyterlab=4.0.9=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.25.2=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libopenblas=0.3.21=h043d6bf_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
more-itertools=10.2.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
multiprocess=0.70.15=pypi_0
nbclient=0.9.0=pypi_0
nbconvert=7.11.0=pypi_0
nbformat=5.9.2=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
networkx=3.2.1=pypi_0
ninja=1.11.1.1=pypi_0
nltk=3.8.1=pypi_0
notebook=7.0.6=pypi_0
notebook-shim=0.2.3=pypi_0
numpy=1.26.2=pypi_0
numpy-base=1.26.0=py310h8a23956_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-ml-py=12.535.161=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
nvitop=1.3.2=pypi_0
openai=1.29.0=pypi_0
openssl=1.1.1w=h7f8727e_0
orjson=3.10.3=pypi_0
overrides=7.4.0=pypi_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.4=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pyhd8ed1ab_0
peft=0.10.0=pypi_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.1.0=pypi_0
pip=23.3.1=pypi_0
platformdirs=4.0.0=pyhd8ed1ab_0
prometheus-client=0.19.0=pypi_0
prompt-toolkit=3.0.41=pyha770c72_0
prompt_toolkit=3.0.41=hd8ed1ab_0
protobuf=5.26.1=pypi_0
psutil=5.9.6=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=14.0.2=pypi_0
pyarrow-hotfix=0.6=pypi_0
pycparser=2.21=pypi_0
pydantic=2.5.3=pypi_0
pydantic-core=2.14.6=pypi_0
pydub=0.25.1=pypi_0
pygments=2.17.0=pyhd8ed1ab_0
pynvml=11.5.0=pypi_0
pyparsing=3.1.2=pypi_0
python=3.10.0=h12debd9_5
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=1.0.1=pypi_0
python-json-logger=2.0.7=pypi_0
python-multipart=0.0.9=pypi_0
python_abi=3.10=2_cp310
pytrie=0.4.0=pyhd8ed1ab_1
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.0=py310h6a678d5_0
qtconsole=5.5.1=pypi_0
qtpy=2.4.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.31.1=pypi_0
regex=2023.10.3=pypi_0
requests=2.31.0=pypi_0
responses=0.18.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.7.0=pypi_0
rpds-py=0.13.2=pypi_0
ruamel-yaml=0.18.6=pypi_0
ruamel-yaml-clib=0.2.8=pypi_0
ruff=0.4.4=pypi_0
s3transfer=0.10.1=pypi_0
safetensors=0.4.3=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.13.0=pypi_0
seaborn=0.13.2=pypi_0
semantic-version=2.10.0=pypi_0
send2trash=1.8.2=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=68.0.0=py310h06a4308_0
shellingham=1.5.4=pypi_0
shtab=1.6.5=pypi_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pypi_0
sortedcontainers=2.4.0=pyhd8ed1ab_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.37.2=pypi_0
sympy=1.12=pypi_0
termcolor=2.4.0=pypi_0
terminado=0.18.0=pypi_0
threadpoolctl=3.2.0=pypi_0
tiktoken=0.6.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.19.1=pypi_0
tomli=2.0.1=pypi_0
tomlkit=0.12.0=pypi_0
toolz=0.12.1=pypi_0
torch=2.1.1=pypi_0
torchaudio=2.1.1=pypi_0
torchvision=0.16.1=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.13.0=pyhd8ed1ab_0
transformers=4.43.1=pypi_0
transformers-stream-generator=0.0.5=pypi_0
triton=2.1.0=pypi_0
trl=0.7.7=pypi_0
typeguard=4.2.1=pypi_0
typer=0.12.3=pypi_0
types-python-dateutil=2.8.19.14=pypi_0
typing-extensions=4.11.0=pypi_0
tyro=0.6.3=pypi_0
tzdata=2023.3=pypi_0
ujson=5.10.0=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.1.0=pypi_0
uvicorn=0.29.0=pypi_0
uvloop=0.19.0=pypi_0
watchfiles=0.21.0=pypi_0
wcwidth=0.2.10=pyhd8ed1ab_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.6.4=pypi_0
websockets=11.0.3=pypi_0
wheel=0.41.2=py310h06a4308_0
widgetsnbextension=4.0.9=pypi_0
xformers=0.0.23.post1=pypi_0
xxhash=3.4.1=pypi_0
xz=5.4.2=h5eee18b_0
yarl=1.9.4=pypi_0
zeromq=4.3.4=h2531618_0
zhconv=1.4.3=pypi_0
zlib=1.2.13=h5eee18b_0
